refactor(bus): extract generic peripheral arms into a cross-vendor factory
Move the 32 cross-vendor / generic Cortex-M peripheral arms (uart, timer,
spi, i2c, rcc, flash, dma, adc, rtc, usb_otg, can, …) out of the central
from_config match into peripherals/generic_factory.rs (try_build). The match
now holds only the YAML descriptor loaders (declarative / strict_ir) and the
unknown-type stub.

from_config delegates: esp32s3 + nrf52 family factories, then the generic
factory (fallible — size/profile parsing), then the descriptor loaders.
parse_profile_or_default / profile_name made pub(crate) for the factory.

bus/mod.rs from_config: -393 lines (−530 with the nrf52 extract). Behaviour
identical: register_coverage_ratchet unchanged across all 15 chips +
firmware_survival (26) + workspace lib + brom_boot_smoke.
